+++ event-sourcing/subscription-manager/src/main/java/uk/gov/justice/services/event/sourcing/subscription/manager/NewSubscriptionManagerDelegate.java
@@ -2,7 +2,9 @@ package uk.gov.justice.services.event.sourcing.subscription.manager;
 
 import static uk.gov.justice.services.event.sourcing.subscription.manager.EventOrderingStatus.EVENT_CORRECTLY_ORDERED;
 
+import uk.gov.justice.services.eventsourcing.util.messaging.EventSourceNameCalculator;
 import uk.gov.justice.services.messaging.JsonEnvelope;
+import uk.gov.justice.services.metrics.micrometer.counters.MicrometerMetricsCounters;
 
 import javax.inject.Inject;
 
@@ -15,7 +17,16 @@ public class NewSubscriptionManagerDelegate {
     @Inject
     private StreamStatusService streamStatusService;
 
+    @Inject
+    private MicrometerMetricsCounters micrometerMetricsCounters;
+
+    @Inject
+    private EventSourceNameCalculator eventSourceNameCalculator;
+
     public void process(final JsonEnvelope incomingJsonEnvelope, final String componentName) {
+
+        final String source = eventSourceNameCalculator.getSource(incomingJsonEnvelope);
+        micrometerMetricsCounters.incrementEventsReceivedCount(source, componentName);
 
         final EventOrderingStatus eventOrderingStatus = streamStatusService.handleStreamStatusUpdates(
                 incomingJsonEnvelope,