+++ event-sourcing/subscription-manager/src/test/java/uk/gov/justice/services/event/sourcing/subscription/manager/NewSubscriptionManagerDelegateTest.java
@@ -8,7 +8,9 @@ import static uk.gov.justice.services.event.sourcing.subscription.manager.EventO
 import static uk.gov.justice.services.event.sourcing.subscription.manager.EventOrderingStatus.EVENT_CORRECTLY_ORDERED;
 import static uk.gov.justice.services.event.sourcing.subscription.manager.EventOrderingStatus.EVENT_OUT_OF_ORDER;
 
+import uk.gov.justice.services.eventsourcing.util.messaging.EventSourceNameCalculator;
 import uk.gov.justice.services.messaging.JsonEnvelope;
+import uk.gov.justice.services.metrics.micrometer.counters.MicrometerMetricsCounters;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -25,6 +27,12 @@ public class NewSubscriptionManagerDelegateTest {
     @Mock
     private StreamStatusService streamStatusService;
 
+    @Mock
+    private MicrometerMetricsCounters micrometerMetricsCounters;
+
+    @Mock
+    private EventSourceNameCalculator eventSourceNameCalculator;
+
     @InjectMocks
     private NewSubscriptionManagerDelegate newSubscriptionManagerDelegate;
 
@@ -33,13 +41,17 @@ public class NewSubscriptionManagerDelegateTest {
 
         final JsonEnvelope incomingJsonEnvelope = mock(JsonEnvelope.class);
         final String componentName = "some-component-name";
+        final String source = "some-source";
 
+        when(eventSourceNameCalculator.getSource(incomingJsonEnvelope)).thenReturn(source);
         when(streamStatusService.handleStreamStatusUpdates(
                 incomingJsonEnvelope,
                 componentName)).thenReturn(EVENT_CORRECTLY_ORDERED);
 
         newSubscriptionManagerDelegate.process(incomingJsonEnvelope, componentName);
 
+        verify(eventSourceNameCalculator).getSource(incomingJsonEnvelope);
+        verify(micrometerMetricsCounters).incrementEventsReceivedCount(source,componentName);
         verify(eventBufferAwareSubscriptionEventProcessor).process(incomingJsonEnvelope, componentName);
     }
 
@@ -48,13 +60,17 @@ public class NewSubscriptionManagerDelegateTest {
 
         final JsonEnvelope incomingJsonEnvelope = mock(JsonEnvelope.class);
         final String componentName = "some-component-name";
+        final String source = "some-source";
 
+        when(eventSourceNameCalculator.getSource(incomingJsonEnvelope)).thenReturn(source);
         when(streamStatusService.handleStreamStatusUpdates(
                 incomingJsonEnvelope,
                 componentName)).thenReturn(EVENT_OUT_OF_ORDER);
 
         newSubscriptionManagerDelegate.process(incomingJsonEnvelope, componentName);
 
+        verify(eventSourceNameCalculator).getSource(incomingJsonEnvelope);
+        verify(micrometerMetricsCounters).incrementEventsReceivedCount(source,componentName);
         verify(eventBufferAwareSubscriptionEventProcessor, never()).process(incomingJsonEnvelope, componentName);
     }
 
@@ -63,13 +79,17 @@ public class NewSubscriptionManagerDelegateTest {
 
         final JsonEnvelope incomingJsonEnvelope = mock(JsonEnvelope.class);
         final String componentName = "some-component-name";
+        final String source = "some-source";
 
+        when(eventSourceNameCalculator.getSource(incomingJsonEnvelope)).thenReturn(source);
         when(streamStatusService.handleStreamStatusUpdates(
                 incomingJsonEnvelope,
                 componentName)).thenReturn(EVENT_ALREADY_PROCESSED);
 
         newSubscriptionManagerDelegate.process(incomingJsonEnvelope, componentName);
 
+        verify(eventSourceNameCalculator).getSource(incomingJsonEnvelope);
+        verify(micrometerMetricsCounters).incrementEventsReceivedCount(source,componentName);
         verify(eventBufferAwareSubscriptionEventProcessor, never()).process(incomingJsonEnvelope, componentName);
     }
 }