+++ CHANGELOG.md
@@ -4,6 +4,8 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 [Semantic Versioning](http://semver.org/).
 
 ### [Unreleased]
+### Changed
+- Added framework.events.received to NewSubscriptionManagerDelegate to count correctly
 
 # [17.103.0-M23] - 2025-07-07
 ### Added